feat(archdispatch): Qwen35Dispatch impls for the expanded AR hooks (Inc 1 Task 1.4b-ii)
Verbatim from the qwen35 AR arm: maybe_evict (→ new physical slot),
maybe_adaptive_downshift, take_prefill_checkpoint, abort_zero_recurrent
(DeltaNet s/conv buffers + KV compact_offset + co-resident Llama KV), sample
(GPU fast path / CPU grammar-mask path), and Qwen35GrammarMatcher::
attractor_detected. Each hook reproduces the arm's field-split borrow of *m
(m.state bundle disjoint from m.eviction/m.kv_adaptive/m.prefill_checkpoints).
Also fixes the 1.4b-i sample() rng_state type u64→u32 to match sampler::sample.
Dead-code; consumed by ar_generate in 1.4b-iii. Builds: daemon example + lib.

+++ crates/hipfire-runtime/src/arch_dispatch.rs
@@ -181,7 +181,7 @@ pub trait ArchDispatch {
         vocab_size: usize,
         ngram_scope: &[u32],
         grammar_mask: Option<&[bool]>,
-        rng_state: &mut u64,
+        rng_state: &mut u32,
     ) -> Result<u32, String> {
         let _ = (gpu, cfg, vocab_size, ngram_scope, grammar_mask, rng_state);
         Err("sample not implemented for this arch".into())
